:truck: Rename typer commands
