+++ data_safe_haven/commands/__init__.py
@@ -1,7 +1,7 @@
-from .typer_admin import admin_command_group
-from .typer_deploy import deploy_command_group
-from .typer_init import initialise_command
-from .typer_teardown import teardown_command_group
+from .admin import admin_command_group
+from .deploy import deploy_command_group
+from .init import initialise_command
+from .teardown import teardown_command_group
 
 __all__ = [
     "admin_command_group",
