+++ data_safe_haven/commands/admin_remove_users.py
