+++ data_safe_haven/commands/deploy_shm.py
