+++ data_safe_haven/commands/teardown_backend.py
