+++ data_safe_haven/commands/teardown_shm.py
